// src/lib/api/models/getApiAppointmentsStatus.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * RealHub API
 * Multi-tenant Real Estate Platform API
 * OpenAPI spec version: v1
 */

export type GetApiAppointmentsStatus = typeof GetApiAppointmentsStatus[keyof typeof GetApiAppointmentsStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetApiAppointmentsStatus = {
  SCHEDULED: 'SCHEDULED',
  CONFIRMED: 'CONFIRMED',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED',
  NO_SHOW: 'NO_SHOW',
} as const;